ml/requirements_fastapi.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.12.0
python-dotenv==1.0.0
openai
exa-py
agno
firecrawl
python-multipart==0.0.6
